Moved code related to the gravity solver into Enzo/gravity directory
